+++ project/config/providers.php
@@ -1,7 +1,0 @@
-<?php
-
-use Project\Providers\ProjectServiceProvider;
-
-return [
-    ProjectServiceProvider::class,
-];