+++ resources/views/components/application-logo.blade.php
@@ -1,1 +1,1 @@
-<img src="{{ asset('brand/logo-64.png') }}" alt="{{ config('app.name') }} logo" {{ $attributes }}>
+<img src="{{ asset('cms/brand/logo-64.png') }}" alt="{{ config('app.name') }} logo" {{ $attributes }}>